requirements.txt
pyspark==3.3.1
PyYAML==6.0
pytest==7.2.0
pandas==2.0.1
faker==18.9.0
numpy==1.24.3
pyarrow==12.0.0  # For Parquet support
datetime